Refactor environment creation for split between train and eval
Major change that refactors the API for create_env.py
create_env.py now creates a train and eval env with the option of the eval env being simply the train env.
All algorithms are adapted to that.
For clean evaluation don't copy the train environment, only for MJX environments that is fine. Per default copy is activated though to save memory.
Also currently the eval env is used for testing.